+++ src/frontend/packages/lib_components/src/common/queries/fetchOne.tsx
@@ -1,0 +1,28 @@
+import { QueryKey, QueryFunctionContext } from 'react-query';
+import { useJwt } from '../../hooks/stores/useJwt';
+
+/**
+ * `react-query` fetcher for individual items from the Marsha API.
+ */
+export const fetchOne = async <T,>({
+  queryKey,
+}: QueryFunctionContext<QueryKey>): Promise<T> => {
+  const jwt = useJwt.getState().jwt;
+  const [name, id, action] = queryKey;
+
+  const actionEndpoint = action ? `${String(action)}/` : '';
+  const endpoint = `/api/${String(name)}/${String(id)}/${actionEndpoint}`;
+
+  const response = await fetch(endpoint, {
+    headers: {
+      'Content-Type': 'application/json',
+      ...(jwt ? { Authorization: `Bearer ${jwt}` } : {}),
+    },
+  });
+
+  if (!response.ok) {
+    throw new Error(`Failed to get ${endpoint}.`);
+  }
+
+  return (await response.json()) as T;
+};